+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Chg #368: Update `QueryBuilder` constructor (@Tigrov)
 - Enh #367: Use `ColumnDefinitionBuilder` to generate table column SQL representation (@Tigrov)
 - Enh #371: Remove `ColumnInterface` (@Tigrov)
+- Enh #372: Rename `ColumnSchemaInterface` to `ColumnInterface` (@Tigrov)
 
 ## 1.2.0 March 21, 2024